added a few comments, moved files around. Updated readme
